moved platform independent definitions to wifly_cmd.h
